refactor: 게시글, 댓글, 대댓글 삭제 권한 설정

+++ src/main/java/com/example/board/comment/service/CommentService.java
@@ -7,6 +7,8 @@ import com.example.board.post.entity.Post;
 import com.example.board.comment.repository.CommentRepository;
 import com.example.board.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,10 +30,14 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public ResponseEntity<?> deleteComment(Long commentId, Long memberId) {
         Comment comment = commentRepository.findByCommentId(commentId)
                 .orElseThrow(() -> new RuntimeException("해당 댓글은 존재하지 않습니다."));
+        if (!comment.getMemberId().equals(memberId)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("삭제 권한이 없습니다.");
+        }
         commentRepository.delete(comment);
+        return ResponseEntity.status(HttpStatus.OK).body("댓글이 삭제되었습니다.");
     }
 
     // 신고 횟수 확인 하고 5가 되면 댓글 삭제
